+++ apps/fluent-tester/src/FluentTester/TestComponents/ContextualMenu/ContextualMenuTest.tsx
@@ -38,13 +38,12 @@ const contextualMenu: React.FunctionComponent<{}> = () => {
     (key) => {
       setLastMenuItemClicked(key);
     },
-    [setLastMenuItemClicked]
+    [setLastMenuItemClicked],
   );
 
   return (
     <View>
       <View style={{ flexDirection: 'row', paddingVertical: 5 }}>
-
         <View style={{ flexDirection: 'column', paddingHorizontal: 5 }}>
           <View style={{ flexDirection: 'row' }}>
             <Text>Should Focus on Mount</Text>
@@ -69,8 +68,8 @@ const contextualMenu: React.FunctionComponent<{}> = () => {
             {lastMenuItemClicked > 0 ? (
               <Text style={{ color: 'blue' }}>{lastMenuItemClicked}</Text>
             ) : (
-                <Text style={{ color: 'blue' }}>none</Text>
-              )}
+              <Text style={{ color: 'blue' }}>none</Text>
+            )}
           </Text>
           <Button content="Press for ContextualMenu" onClick={toggleShowContextualMenu} componentRef={stdBtnRef} />
         </View>
@@ -96,23 +95,23 @@ const contextualMenu: React.FunctionComponent<{}> = () => {
       )}
     </View>
   );
-}
+};
 
 const nestedContextualMenu: React.FunctionComponent<{}> = () => {
   const testImage = require('../Button/icon_24x24.png');
   const testTtf = require('../Button/Font Awesome 5 Free-Solid-900.otf');
 
-    const fontProps: FontIconProps = {
-      fontFamily: `Font Awesome 5 Free`,
-      fontSrcFile: testTtf,
-      codepoint: 0xf083,
-      fontSize: 16,
-    };
+  const fontProps: FontIconProps = {
+    fontFamily: `Font Awesome 5 Free`,
+    fontSrcFile: testTtf,
+    codepoint: 0xf083,
+    fontSize: 16,
+  };
 
-    const svgProps: SvgIconProps = {
-      src: TestSvg,
-      viewBox: '0 0 500 500',
-    };
+  const svgProps: SvgIconProps = {
+    src: TestSvg,
+    viewBox: '0 0 500 500',
+  };
 
   const stdBtnRef = React.useRef(null);
 
@@ -157,17 +156,13 @@ const nestedContextualMenu: React.FunctionComponent<{}> = () => {
     setShowSubmenu(false);
   }, [setShowSubmenu]);
 
-  const onClick = React.useCallback(
-    () => {
-      console.log('submenu item clicked');
-    }, []
-  );
-
+  const onClick = React.useCallback(() => {
+    console.log('submenu item clicked');
+  }, []);
 
   return (
     <View>
       <View style={{ flexDirection: 'row', paddingVertical: 5 }}>
-
         <View style={{ flexDirection: 'column', paddingHorizontal: 5 }}>
           <View style={{ flexDirection: 'row' }}>
             <Text>Should Focus on Mount</Text>
@@ -209,15 +204,15 @@ const nestedContextualMenu: React.FunctionComponent<{}> = () => {
           <ContextualMenuItem icon={testImage} text="Menu item with png Icon" itemKey="1" />
           <ContextualMenuItem icon={{ fontSource: fontProps, color: 'blue' }} text="Menu item with font icon" itemKey="2" />
           <ContextualMenuItem text="Disabled Menu Item" itemKey="3" disabled />
-          <SubmenuItem text="Nested Menu" itemKey="4" onHoverIn={toggleShowSubmenu} componentRef={stdMenuItemRef} />
+          <SubmenuItem icon={{ svgSource: svgProps, width: 12, height: 12 }} text="Nested Menu" itemKey="4" onHoverIn={toggleShowSubmenu} componentRef={stdMenuItemRef} />
           {showSubmenu && (
-            <Submenu
-              target={stdMenuItemRef}
-              onDismiss={onDismissSubmenu}
-              onShow={onShowSubmenu}
-              setShowMenu={toggleShowSubmenu}
-            >
-              <ContextualMenuItem icon={{ svgSource: svgProps, width: 20, height: 20, color: 'red' }} text="SubmenuItem svg icon" itemKey="4" onClick={onClick} />
+            <Submenu target={stdMenuItemRef} onDismiss={onDismissSubmenu} onShow={onShowSubmenu} setShowMenu={toggleShowSubmenu}>
+              <ContextualMenuItem
+                icon={{ svgSource: svgProps, width: 12, height: 12 }}
+                text="SubmenuItem svg icon"
+                itemKey="4"
+                onClick={onClick}
+              />
               <ContextualMenuItem text="SubmenuItem 2" itemKey="2" />
               <ContextualMenuItem text="Disabled Menu Item" itemKey="3" disabled />
             </Submenu>
@@ -227,8 +222,7 @@ const nestedContextualMenu: React.FunctionComponent<{}> = () => {
       )}
     </View>
   );
-
-}
+};
 
 const contextualMenuSections: TestSection[] = [
   {
@@ -239,7 +233,7 @@ const contextualMenuSections: TestSection[] = [
   {
     name: 'Nested ContextualMenu',
     component: nestedContextualMenu,
-  }
+  },
 ];
 
 export const ContextualMenuTest: React.FunctionComponent<{}> = () => {