+++ apps/fluent-tester/src/FluentTester/testPages.ios.ts
@@ -6,6 +6,7 @@ import { ExperimentalButtonTest, HOMEPAGE_BUTTON_BUTTONEXPERIMENTAL } from './Te
 import { HOMEPAGE_ICON_BUTTON, IconTest } from './TestComponents/Icon';
 import { HOMEPAGE_LINK_BUTTON, LinkTest } from './TestComponents/Link';
 import { NativeButtonTest, HOMEPAGE_NATIVEBUTTON_BUTTON } from './TestComponents/NativeButton';
+import { NativeDatePickerTest, HOMEPAGE_NATIVEDATEPICKER_BUTTON } from './TestComponents/NativeDatePicker';
 import { HOMEPAGE_PERSONA_BUTTON, PersonaTest } from './TestComponents/Persona';
 import { HOMEPAGE_PERSONACOIN_BUTTON, PersonaCoinTest } from './TestComponents/PersonaCoin';
 import { HOMEPAGE_PRESSABLE_BUTTON, PressableTest } from './TestComponents/Pressable';
@@ -43,6 +44,11 @@ export const tests: TestDescription[] = [
     name: 'Native Button Test',
     component: NativeButtonTest,
     testPage: HOMEPAGE_NATIVEBUTTON_BUTTON,
+  },
+  {
+    name: 'Native Date Picker',
+    component: NativeDatePickerTest,
+    testPage: HOMEPAGE_NATIVEDATEPICKER_BUTTON,
   },
   {
     name: 'Experimental Button',