+++ packages/components/Checkbox/src/Checkbox.types.ts
@@ -5,6 +5,7 @@ import { IRenderData } from '@uifabricshared/foundation-composable';
 import { ITextProps } from '@fluentui-react-native/text';
 import { FontTokens, IForegroundColorTokens, IBackgroundColorTokens, IBorderTokens } from '@fluentui-react-native/tokens';
 import type { IViewProps } from '@fluentui-react-native/adapters';
+import { IconProps } from '@fluentui-react-native/icon';
 import { ColorValue } from 'react-native';
 
 export const checkboxName = 'Checkbox';
@@ -88,7 +89,8 @@ export interface ICheckboxTokens extends FontTokens, IForegroundColorTokens, IBa
 export interface ICheckboxSlotProps {
   root: React.PropsWithRef<IViewProps>;
   checkbox: IViewProps;
-  checkmark: ITextProps;
+  checkmark?: ITextProps;
+  checkmarkIcon?: IconProps;
   content: ITextProps;
 }