+++ packages/components/Button/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/button
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Wed, 23 Jun 2021 18:53:56 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.4
+
+Wed, 23 Jun 2021 18:53:56 GMT
+
+### Patches
+
+- Add new PersonaCoinFluentColor prop (ruaraki@microsoft.com)
+
+## 0.17.2
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to components, run (ruaraki@microsoft.com)
+
+## 0.17.1
+
+Wed, 16 Jun 2021 04:20:49 GMT
+
+### Patches
+
+- Remove apple theme native module (sanajmi@microsoft.com)
 
 ## 0.17.0