+++ packages/components/ContextualMenu/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/contextual-menu
 
-This log was last generated on Wed, 19 May 2021 07:04:54 GMT and should not be manually modified.
+This log was last generated on Fri, 25 Jun 2021 19:39:16 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.0
+
+Fri, 25 Jun 2021 19:39:16 GMT
+
+### Minor changes
+
+- Add chevron to submenu (lehon@microsoft.com)
+
+## 0.8.4
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to components, run (ruaraki@microsoft.com)
 
 ## 0.8.0