+++ packages/theming/theme-types/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/theme-types
 
-This log was last generated on Tue, 15 Jun 2021 00:55:36 GMT and should not be manually modified.
+This log was last generated on Fri, 25 Jun 2021 06:21:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.0
+
+Fri, 25 Jun 2021 06:21:11 GMT
+
+### Minor changes
+
+- Add activity ring to PersonaCoin (tamasane@gmail.com)
+
+## 0.7.2
+
+Mon, 21 Jun 2021 11:19:28 GMT
+
+### Patches
+
+- Add checkmarkColor to apple theme (tamasane@gmail.com)
+
+## 0.7.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to libraries core and theming, run (ruaraki@microsoft.com)
 
 ## 0.7.0